fix requested chagnes
